+ Store ingredients for app to use in RecipeViewController in an IngredientList (returned from JSON parsing method) + Display ingredient quantity, measurement, and name/brand after 'Add' steps (compare ingredient_id) + fixed ingredients.json + added full rum and coke recipe to recipe.json (currently displays for all drinks)

+++ barbot/barbot/Step.swift
@@ -13,7 +13,7 @@ struct Step: Decodable {
     let step_number: Int
     let type: String
     let ingredient_id: String?
-    let quantity: Int?
+    let quantity: Double?
     let measurement: String?
     
     init?(json: JSON) {
